+++ src/App.tsx
@@ -3,12 +3,13 @@ import RecipeTile from "./components/recipes/RecipeTile";
 import mashedPotatoes from "./interfaces/examples/Recipe";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import './App.css';
-import RecipePage from './components/recipes/RecipePage';
+import RecipePage from './pages/recipes/RecipePage';
+import RecipesPage from './pages/recipes/RecipesPage';
 
 const router = createBrowserRouter([
   {
     path: "/",
-    element: <RecipeTile recipe={mashedPotatoes} />
+    element: <RecipesPage />
   },
   {
     path: "recipe/:recipeId",
