--- src/components/recipes/RecipePage.tsx
@@ -1,28 +1,0 @@
-import { useEffect, useState } from "react";
-import { useParams } from "react-router-dom";
-import { recipes } from "../../interfaces/examples/Recipe";
-import { Recipe } from "../../interfaces/recipes/Recipe";
-import Ingredients from "./Ingredients";
-import Equipment from "./Equipment";
-import Instructions from "./Instructions";
-
-export default function RecipePage() {
-  const { recipeId } = useParams();
-  const [recipe, setRecipe] = useState<Recipe>();
-
-  useEffect(() => {
-    setRecipe(recipes.find(r => r.id == recipeId));
-  }, [])
-
-  return(
-    <div>
-      <h1>{recipe?.name}</h1>
-      <p>Difficulty: {recipe?.difficulty}</p>
-      <p>Cuisine: {recipe?.cuisine}</p>
-      <p></p>
-      <Ingredients ingredients={recipe?.ingredients!} />
-      <Equipment equipment={recipe?.equipment!} />
-      <Instructions instructions={recipe?.instructions!} />
-    </div>
-  );
-}
