+++ src/interfaces/examples/ExampleRecipeList.ts
@@ -1,0 +1,3 @@
+import mashedPotatoes from "./Recipe";
+
+export default Array(30).fill(mashedPotatoes);
